requirements.txt
certifi==2016.9.26
future==0.16.0
python-telegram-bot==5.3.0
urllib3==1.19.1
